+++ app/app-entry/entry.jsx
@@ -12,6 +12,8 @@ import { PageLifecycle } from './PageLifecycle';
 import { moduleSupport } from './module-support';
 import { AppState } from './AppStateProvider';
 import { WorkspaceHooksContextProvider } from 'workspace-hooks/index';
+import { WatchWorkspace } from './WatchWorkspace';
+import { WatchUI } from './WatchUI';
 
 if (typeof window !== undefined) {
   window.Sentry?.onLoad(function () {
@@ -61,6 +63,8 @@ ReactDOM.render(
       <Router>
         <UIManager>
           <WorkspaceHooksContextProvider>
+            <WatchWorkspace />
+            <WatchUI />
             <EditorManager bangleIOContext={bangleIOContext}>
               <App />
             </EditorManager>